+++ src/redux/modules/PostSlice.js
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import api from "../../server/api";
+import { postAPI } from "../../server/api";
 
 const initialState = {
   post: [],
@@ -39,6 +40,22 @@ export const getInfiniteList = createAsyncThunk(
   }
 );
 
+// post요청
+export const _addPost = createAsyncThunk(
+  "post/posts",
+  async (payload, thunkAPI) => {
+    try {
+      console.log("payload", payload);
+      const data = await postAPI.writePost(payload);
+      console.log(data);
+      return thunkAPI.fulfillWithValue(data.data);
+    } catch (error) {
+      console.log(error);
+      return thunkAPI.rejectWithValue(error);
+    }
+  }
+);
+
 const post = createSlice({
   name: "post",
   initialState,
@@ -65,13 +82,25 @@ const post = createSlice({
     [getInfiniteList.pending]: (state) => {
       console.log("확인"); // 네트워크 요청이 시작되면 로딩상태를 true로 변경합니다.
     },
-
     [getInfiniteList.fulfilled]: (state, action) => {
       state.isLoading = false;
       state.post = [...state.post, ...action.payload.content];
       state.last = action.payload.last;
       state.totalPage = action.payload.totalPages;
       console.log(action.payload.content);
+    },
+
+    [_addPost.pending]: (state) => {
+      state.isLoading = true;
+    },
+    [_addPost.fulfilled]: (state, action) => {
+      state.isLoading = false;
+      state.posts = action.payload;
+      console.log("_addPost", action.payload);
+    },
+    [_addPost.rejected]: (state, action) => {
+      state.isLoading = false;
+      state.error = action.payload;
     },
   },
 });
